# Improve unpickling performance when reading from memory buffer