--- a/dotnet/Razorvine.Pyrolite/Benchmarks/Benchmarks.csproj
+++ b/dotnet/Razorvine.Pyrolite/Benchmarks/Benchmarks.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp2.0</TargetFramework>
-    <LangVersion>7</LangVersion>
+    <LangVersion>7.3</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>