--- a/dotnet/Razorvine.Pyrolite/Benchmarks/Program.cs
+++ b/dotnet/Razorvine.Pyrolite/Benchmarks/Program.cs
@@ -1,0 +1,17 @@
+﻿using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Diagnosers;
+using BenchmarkDotNet.Jobs;
+using BenchmarkDotNet.Running;
+
+namespace Benchmarks
+{
+    class Program
+    {
+        static void Main(string[] args) 
+            => BenchmarkSwitcher
+                .FromAssembly(typeof(Program).Assembly)
+                .Run(args, DefaultConfig.Instance
+                    .With(MemoryDiagnoser.Default)
+                    .With(Job.ShortRun));
+    }
+}