--- a/dotnet/Razorvine.Pyrolite/Pyrolite/Pickle/Unpickler.cs
+++ b/dotnet/Razorvine.Pyrolite/Pyrolite/Pickle/Unpickler.cs
@@ -79,9 +79,20 @@ public class Unpickler : IDisposable {
     /// Read a pickled object representation from the given pickle data bytes.
     /// </summary>
     /// <param name="pickledata">Serialized pickle data.</param>
-    /// <param name="stackCapacity">Optional parameter that suggests the initial capacity of stack. The default value is 4.</param>
     /// <returns>the reconstituted object hierarchy specified in the memory buffer.</returns>
-    public object loads(byte[] pickledata, int stackCapacity = UnpickleStack.DefaultCapacity) {
+    public object loads(byte[] pickledata) {
+        stack = new UnpickleStack();
+        var unpickler = new Unpickler<ArrayReader>(new ArrayReader(pickledata), memo, stack, this);
+        return unpickler.Load();
+    }
+
+    /// <summary>
+    /// Read a pickled object representation from the given pickle data bytes.
+    /// </summary>
+    /// <param name="pickledata">Serialized pickle data.</param>
+    /// <param name="stackCapacity">Initial capacity of the UnpickleStack.</param>
+    /// <returns>the reconstituted object hierarchy specified in the memory buffer.</returns>
+    public object loads(byte[] pickledata, int stackCapacity) {
         stack = new UnpickleStack(stackCapacity);
         var unpickler = new Unpickler<ArrayReader>(new ArrayReader(pickledata), memo, stack, this);
         return unpickler.Load();