--- a/dotnet/Razorvine.Pyrolite/Razorvine.Pyrolite.sln
+++ b/dotnet/Razorvine.Pyrolite/Razorvine.Pyrolite.sln
@@ -14,6 +14,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NamingExample", "NamingExam
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{59C33EFD-F2D7-4FC7-89FF-ACF808AE1109}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Benchmarks", "Benchmarks\Benchmarks.csproj", "{9E88F4F5-A18E-4ACC-8864-90379CE095DF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -48,5 +50,9 @@ Global
 		{59C33EFD-F2D7-4FC7-89FF-ACF808AE1109}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{59C33EFD-F2D7-4FC7-89FF-ACF808AE1109}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{59C33EFD-F2D7-4FC7-89FF-ACF808AE1109}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9E88F4F5-A18E-4ACC-8864-90379CE095DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9E88F4F5-A18E-4ACC-8864-90379CE095DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9E88F4F5-A18E-4ACC-8864-90379CE095DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9E88F4F5-A18E-4ACC-8864-90379CE095DF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal