When the Unpickler reads from a `Stream` it has non-trivial overhead compared to when reading from a memory buffer. This PR improves unpickling performance for that case.

The difference
- `stream.ReadByte` with an additional check (if the byte is not `-1`) is just `array[position++]`.
-  `stream.Read(buffer, 0, size)` is just `array.Slice(0, size)`. This helps to avoid calling `Array.Copy` and copying the data more than needed.

Perf trick:  the types that implement `IInputReader` are structs to enable for full interface method devirtualization and inlining. You can read about it [here](https://adamsitnik.com/Value-Types-vs-Reference-Types/#how-to-avoid-boxing-with-value-types-that-implement-interfaces). I have added the necessary comment in the code that explains that.

The gains are quite big. From 20% to 2 times faster when unpickling from array. Unpickling from Stream has not been regressed. I added micro-benchmarks to measure and ensure this.

|             Method | Count |        WAS | IS netcoreapp2.0 | IS netcoreapp3.0 |
|------------------- |------ |-----------:|-----------------:|-----------------:|
|  DoublesFromStream |   100 |   7.151 us |         7.174 us |         5.594 us |
| IntegersFromStream |   100 |   4.859 us |         4.427 us |         4.240 us |
| BooleansFromStream |   100 |   2.183 us |         1.917 us |         1.824 us |
|  StringsFromStream |   100 |  14.648 us |        12.766 us |        10.439 us |
|   DoublesFromArray |   100 |   7.246 us |         5.020 us |         3.912 us |
|  IntegersFromArray |   100 |   4.802 us |         3.897 us |         3.822 us |
|  BooleansFromArray |   100 |   2.154 us |         1.653 us |         1.626 us |
|   StringsFromArray |   100 |  13.254 us |         9.353 us |         7.298 us |
|  DoublesFromStream |  1000 |  61.817 us |        60.708 us |        48.213 us |
| IntegersFromStream |  1000 |  52.808 us |        50.321 us |        39.714 us |
| BooleansFromStream |  1000 |  18.297 us |        15.809 us |        14.560 us |
|  StringsFromStream |  1000 | 151.480 us |       142.836 us |       112.647 us |
|   DoublesFromArray |  1000 |  **60.844** us |        41.348 us |        **31.345** us |
|  IntegersFromArray |  1000 |  **52.648** us |        40.204 us |        **31.525** us |
|  BooleansFromArray |  1000 |  **18.346** us |        13.245 us |        **13.369** us |
|   StringsFromArray |  1000 | **153.078** us |        99.069 us |        **72.801** us |

/cc @stephentoub, @imback82, @eerhardt, @rapoth