@irmen 2019-05-04
I have been on holiday so haven't been able to follow this PR until now. Will have a look at all the stuff that was talked about tomorrow, I think