@adamsitnik 2019-05-05
@irmen I have pushed all the code fixes, the code is ready for review.