@irmen 2019-05-06
I hope to be able to review all this in the following few days, but it's clear to me that you guys are a lot more knowledgeable about these low level optimizations than I am. So I don't think I'll be able to comment on the function, but rather only on the form (if any)