@irmen 2019-05-08
thanks a lot for these PRs and it's exiting that it makes Pyrolite a key library in .NET Spark