@adamsitnik 2019-05-08
@irmen thank you for the awesome library!

BTW please don't release a new NuGet package yet, tomorrow I plan to finish a PR that improves the Pickler performance.

Preview:

|              Method | Count |   Mean WAS |    Mean IS | WAS Alloc | IS Allocated |
|-------------------- |------ |-----------:|-----------:|----------:|-------------:|
|  DoublesToByteArray |   100 |   3.138 us |   0.911 us |    3184 B |        304 B |
|  StringsToByteArray |   100 |  22.235 us |  16.828 us |   16424 B |      10280 B |
|  DoublesToByteArray |  1000 |  31.298 us |   7.269 us |   42105 B |        304 B |
|  StringsToByteArray |  1000 | 234.243 us | 169.381 us |  180100 B |     102306 B |