@irmen 2019-05-08
@adamsitnik  Those are incredible results!  I'm very curious what wizardry you're performing here

I'm just in the process right now of spicing up the readme.